% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interimData1}
\alias{interimData1}
\title{Interim enrollment and event data before enrollment completion}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 224 rows and 9 columns.
}
\usage{
interimData1
}
\description{
A data frame with 225 rows and 9 columns:
\describe{
  \item{\code{trialsdt}}{The trial start date}
  \item{\code{usubjid}}{The unique subject ID}
  \item{\code{randdt}}{The randomization date}
  \item{\code{treatment}}{The treatment group number}
  \item{\code{treatment_description}}{Description of the treatment group}
  \item{\code{time}}{The day of event or censoring since randomization}
  \item{\code{event}}{The event indicator: 1 for event, 0 for non-event}
  \item{\code{dropout}}{The dropout indicator: 1 for dropout,
  0 for non-dropout}
  \item{\code{cutoffdt}}{The cutoff date}
}
For ongoing subjects, both \code{event} and \code{dropout} are equal to 0.
}
\keyword{datasets}
