% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_Chen_log_logistic.R
\name{loglik_Chen_log_logistic}
\alias{loglik_Chen_log_logistic}
\title{Log-likelihood function for log-logistic distribution with cured population}
\usage{
loglik_Chen_log_logistic(par, df)
}
\arguments{
\item{par}{a vector with three elements, where the first element denotes the logistic of the proportion
of the cured population, and the second element and the third element denote
the log of shape and the log of the scale parameter of the log-logistic distribution.}

\item{df}{The subject-level event data, including \code{time}
and \code{event}.}
}
\value{
The negative value of the log-likelihood function given parameter
\code{par} and the dataset \code{df}
}
\description{
Provide log-likelihood function for log-logistic distribution with cured population
}
\references{
\itemize{
\item Chen, Tai-Tsang. "Predicting analysis times in randomized clinical trials with cancer immunotherapy." 
BMC medical research methodology 16.1 (2016): 1-10.
}
}
