% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_Chen_weibull2.R
\name{loglik_Chen_weibull2}
\alias{loglik_Chen_weibull2}
\title{Log-likelihood function for Weibull distribution with cured population}
\usage{
loglik_Chen_weibull2(par, df, p)
}
\arguments{
\item{par}{a vector with two elements denoting
the log of the shape and log of the scale parameter of the Weibull distribution.}

\item{df}{The subject-level event data, including \code{time}
and \code{event}.}

\item{p}{cure rate}
}
\value{
The negative value of the log-likelihood function given parameter
\code{par} and the dataset \code{df}
}
\description{
Provide log-likelihood function for Weibull distribution with cured population.
}
\references{
\itemize{
\item Chen, Tai-Tsang. "Predicting analysis times in randomized clinical trials with cancer immunotherapy." 
BMC medical research methodology 16.1 (2016): 1-10.
}
}
