% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smed.R
\name{smed}
\alias{smed}
\title{Function to output summary statistics from survfit function output}
\usage{
smed(x)
}
\arguments{
\item{x}{an object returned by "survfit"}
}
\value{
"smed" returns a matrix of 5 columns of
}
\description{
Provide  summary statistics from survfit function output
}
\examples{
n <- 500
event <- runif(n,1, 5)
osc<-1*(event<=4)
os <- pmin(event,4)

fit1<-survival::survfit(survival::Surv(os,osc)~1)
smed(fit1)

}
