% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP_Chen_weibull.R
\name{SP_Chen_weibull}
\alias{SP_Chen_weibull}
\title{Survival probability function of the Weibull distribution with cured population}
\usage{
SP_Chen_weibull(par, t)
}
\arguments{
\item{par}{a vector with three elements, where the first element denotes the logistic of the proportion
of the cured population, and the second element and the third element denote
the log of the shape and log of the scale parameter of the Weibull distribution.}

\item{t}{a vector with non-negative elements.}
}
\value{
Survival probability of the Weibull distribution with cured population given parameters \code{par} at times \code{t}.
}
\description{
Calculate survival probability of the Weibull distribution with cured population
}
\references{
\itemize{
\item Chen, Tai-Tsang. "Predicting analysis times in randomized clinical trials with cancer immunotherapy." 
BMC medical research methodology 16.1 (2016): 1-10.
}
}
